requirements.txt
PySide6==6.10.2
websocket-client==1.8.0
